pyproject.toml
[project]
name = "de-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "dlt>=1.23.0",
    "kafka-python>=2.3.0",
    "pandas>=3.0.1",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
